requirements.txt
requests>=2.31.0
tqdm>=4.66.0
openai>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
playwright>=1.40.0
